I've refactored the integration test helpers and updated the README.
Here's a summary of what I did:
- I moved `loadAllClientScripts` to `tests/client/shared/integrationTestSetup.js`.
- I renamed `loadAllClientScripts` to `setupIntegrationTestEnvironment` for better clarity.
- I updated all integration tests to use the new function and path.
- I cleaned up `tests/client/shared/testHelpers.js` by removing the old function.
- I updated `README.md` to reflect these changes, clarifying the use of `setupIntegrationTestEnvironment` for integration tests and `loadClientScript` for more isolated unit tests of client scripts.
- I ensured all tests pass after these changes.
